+++ packages/shared/src/task/video.ts
@@ -39,17 +39,20 @@ export const getFfmpegPath = () => {
   let ffmpegPath = config.ffmpegPath;
   let ffprobePath = config.ffprobePath;
   let mesioPath = config.mesioPath;
+  let bililiveRecorderPath = config.bililiveRecorderPath;
   if (!config.customExecPath) {
     const globalConfig = container.resolve<GlobalConfig>("globalConfig");
     ffmpegPath = globalConfig.defaultFfmpegPath;
     ffprobePath = globalConfig.defaultFfprobePath;
     mesioPath = globalConfig.defaultMesioPath;
+    bililiveRecorderPath = globalConfig.defaultBililiveRecorderPath;
   }
 
   return {
     ffmpegPath,
     ffprobePath,
     mesioPath,
+    bililiveRecorderPath,
   };
 };