+++ packages/BilibiliRecorder/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next
 
 - 新增`debugLevel`参数，支持`none`、`basic`、`verbose`
+- 触发标题黑名单设定额外状态
 
 # 1.7.1